add new lib und entity system

+++ source/code/intern/src/core/core_id_manager.h
@@ -13,7 +13,7 @@ namespace Core
     {
         public:
 
-            using BID = unsigned int;
+            using BID = size_t;
 
         public:
